FillForm: Move main body of work into FormImporter.
