+++ fix_throws_javadoc.sed
@@ -17,3 +17,6 @@ s/^\(  *\*  *@throws UnsupportedFontException\).*$/\1 the font is not supported/
 s/^\(  *\*  *@throws ClassNotFoundException\).*$/\1 a class definition could not be found/
 s/^\(  *\*  *@throws URISyntaxException\).*$/\1 a string could not be parsed as a URI reference/
 s/^\(  *\*  *@throws PDFInvalidXMLException\).*$/\1 The XML passed to the method either directly or indirectly is invalid/
+s/^\(  *\*  *@throws SAXException\).*$/\1 basic error or warning information from either the XML parser or the application/
+s/^\(  *\*  *@throws TransformerException\).*$/\1 an exceptional condition that occured during the transformation process/
+s/^\(  *\*  *@throws ParserConfigurationException\).*$/\1 a serious configuration error/
