+++ src/main/java/com/datalogics/pdf/samples/forms/FillForm.java
@@ -4,53 +4,22 @@
 
 package com.datalogics.pdf.samples.forms;
 
-import com.adobe.internal.io.ByteReader;
-import com.adobe.internal.io.InputStreamByteReader;
 import com.adobe.pdfjt.Version;
-import com.adobe.pdfjt.core.exceptions.PDFConfigurationException;
-import com.adobe.pdfjt.core.exceptions.PDFFontException;
-import com.adobe.pdfjt.core.exceptions.PDFIOException;
-import com.adobe.pdfjt.core.exceptions.PDFInvalidDocumentException;
-import com.adobe.pdfjt.core.exceptions.PDFInvalidParameterException;
-import com.adobe.pdfjt.core.exceptions.PDFInvalidXMLException;
-import com.adobe.pdfjt.core.exceptions.PDFSecurityException;
-import com.adobe.pdfjt.core.exceptions.PDFUnableToCompleteOperationException;
 import com.adobe.pdfjt.core.license.LicenseManager;
 import com.adobe.pdfjt.pdf.document.PDFDocument;
 import com.adobe.pdfjt.pdf.document.PDFDocument.PDFDocumentType;
-import com.adobe.pdfjt.services.ap.AppearanceService;
-import com.adobe.pdfjt.services.fdf.FDFDocument;
-import com.adobe.pdfjt.services.fdf.FDFService;
-import com.adobe.pdfjt.services.forms.FormFieldService;
 import com.adobe.pdfjt.services.xfa.XFAService;
-import com.adobe.pdfjt.services.xfa.XFAService.XFAElement;
-import com.adobe.pdfjt.services.xfdf.XFDFService;
 
-import com.datalogics.pdf.document.DocumentHelper;
 import com.datalogics.pdf.samples.util.DocumentUtils;
 import com.datalogics.pdf.samples.util.IoUtils;
 
 import org.apache.maven.artifact.versioning.DefaultArtifactVersion;
-import org.w3c.dom.Document;
-import org.w3c.dom.Element;
-import org.w3c.dom.Node;
 
-import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
-import java.net.URISyntaxException;
 import java.net.URL;
 import java.util.Locale;
 import java.util.Properties;
-
-import javax.xml.parsers.DocumentBuilder;
-import javax.xml.parsers.DocumentBuilderFactory;
-import javax.xml.transform.Result;
-import javax.xml.transform.Source;
-import javax.xml.transform.Transformer;
-import javax.xml.transform.TransformerFactory;
-import javax.xml.transform.dom.DOMSource;
-import javax.xml.transform.stream.StreamResult;
 
 /**
  * This sample will demonstrate how to fill different types of PDF forms. For Acroforms, FDF and XFDF form data formats
@@ -180,17 +149,7 @@ public final class FillForm {
      */
     public static void fillAcroformFdf(final PDFDocument pdfDocument, final URL inputDataUrl, final URL outputUrl)
                     throws Exception {
-
-        // Open the input form data file.
-        final InputStream formStream = inputDataUrl.openStream();
-        final ByteReader formByteReader = new InputStreamByteReader(formStream);
-        final FDFDocument fdfDocument = FDFDocument.newInstance(formByteReader);
-
-        // Use the FDFService to get the form data into the PDF.
-        final FDFService fdfService = new FDFService(pdfDocument);
-        fdfService.importForm(fdfDocument);
-
-        finishAndSaveForm(pdfDocument, outputUrl);
+        FormImporter.fillAcroformFdf(pdfDocument, inputDataUrl, outputUrl);
     }
 
     /**
@@ -203,51 +162,7 @@ public final class FillForm {
      */
     public static void fillAcroformXfdf(final PDFDocument pdfDocument, final URL inputDataUrl, final URL outputUrl)
                     throws Exception {
-
-        // If this is XFDF form data, fill the form using the XFDFService, which uses a slightly different
-        // process than the FDFService. Just get the data file into an InputStream, then import the data into the PDF
-        // document.
-        final InputStream formStream = inputDataUrl.openStream();
-        XFDFService.importFormData(pdfDocument, formStream);
-
-        // Run calculations on the AcroForm...only required before PDFJT 4.5.0
-        finishAndSaveForm(pdfDocument, outputUrl);
-    }
-
-    /**
-     * Run scripts, generate appearances, and save document.
-     *
-     * @param pdfDocument the document to complete and save
-     * @param outputUrl the URL to which to save the file
-     *
-     * @throws IOException an I/O operation failed or was interrupted
-     * @throws PDFUnableToCompleteOperationException the operation was unable to be completed
-     * @throws PDFInvalidDocumentException a general problem with the PDF document, which may now be in an invalid state
-     * @throws PDFIOException there was an error reading or writing a PDF file or temporary caches
-     * @throws PDFSecurityException some general security issue occurred during the processing of the request
-     * @throws PDFInvalidParameterException one or more of the parameters passed to a method is invalid
-     * @throws PDFConfigurationException there was a system problem configuring PDF support
-     * @throws PDFInvalidXMLException The XML passed to the method either directly or indirectly is invalid
-     * @throws PDFFontException there was an error in the font set or an individual font
-     * @throws URISyntaxException a string could not be parsed as a URI reference
-     */
-    private static void finishAndSaveForm(final PDFDocument pdfDocument, final URL outputUrl)
-                    throws IOException, PDFInvalidDocumentException, PDFSecurityException, PDFIOException,
-                    PDFInvalidParameterException, PDFInvalidXMLException, PDFConfigurationException,
-                    PDFUnableToCompleteOperationException, PDFFontException, URISyntaxException {
-        if (pdfjtIsBeforeVersion4()) {
-            // Run calculations scripts on the AcroForm...only required before PDFJT 4.5.0
-            FormFieldService.getAcroFormFieldManager(pdfDocument).runCalculateScripts();
-
-            // Run format scripts on the AcroForm...only required before PDFJT 4.7.0
-            FormFieldService.getAcroFormFieldManager(pdfDocument).runFormatScripts();
-        }
-
-        // And generate appearances.
-        AppearanceService.generateAppearances(pdfDocument, null, null);
-
-        // Save the file.
-        DocumentHelper.saveFullAndClose(pdfDocument, outputUrl.toURI().getPath());
+        FormImporter.fillAcroformXfdf(pdfDocument, inputDataUrl, outputUrl);
     }
 
     /**
@@ -260,112 +175,7 @@ public final class FillForm {
      */
     public static void fillXfa(final PDFDocument pdfDocument, final URL inputDataUrl, final URL outputUrl)
                     throws Exception {
-        // Check to see if the input data file is properly formatted for the XFAService to use. The two outermost
-        // elements should be <xfa:datasets><xfa:data>; if they're not, make it so.
-        if (!hasXfaRootHeader(inputDataUrl)) {
-            addXfaRootHeader(new File(inputDataUrl.toURI()));
-        }
-        if (!hasXfaChildHeader(inputDataUrl)) {
-            addXfaChildHeader(new File(inputDataUrl.toURI()));
-        }
-
-        // Start by getting the form data into an InputStream.
-        final InputStream formStream = inputDataUrl.openStream();
-
-        // If we have an XML file with the proper header, use the XFAService to get the data into the PDF.
-        XFAService.importElement(pdfDocument, XFAElement.DATASETS, formStream);
-
-        // Just save the file. Generating appearances and running calculations aren't supported for XFA forms, so
-        // there's no need to try it.
-        DocumentHelper.saveFullAndClose(pdfDocument, outputUrl.toURI().getPath());
-    }
-
-    /**
-     * Determine if the first element is an xfadataset root element as expected.
-     *
-     * @param inputDataUrl a file containing XML form data
-     * @return a boolean indicating the presence of XFA header tags.
-     * @throws Exception a general exception was thrown
-     */
-    private static boolean hasXfaRootHeader(final URL inputDataUrl) throws Exception {
-
-        final DocumentBuilderFactory factory = DocumentBuilderFactory.newInstance();
-        final DocumentBuilder builder = factory.newDocumentBuilder();
-
-        final Document xmlDoc = builder.parse(inputDataUrl.openStream());
-        final String xmlRootName = xmlDoc.getDocumentElement().getNodeName();
-
-        return xmlRootName.equals(XFA_DATA_ROOT_NODE);
-    }
-
-    /**
-     * Determine if the first child element is an xfadata element as expected.
-     *
-     * @param inputDataUrl a file containing XML form data
-     * @return a boolean indicating the presence of XFA header tags.
-     * @throws Exception a general exception was thrown
-     */
-    private static boolean hasXfaChildHeader(final URL inputDataUrl) throws Exception {
-
-        final DocumentBuilderFactory factory = DocumentBuilderFactory.newInstance();
-        final DocumentBuilder builder = factory.newDocumentBuilder();
-
-        final Document xmlDoc = builder.parse(inputDataUrl.openStream());
-        final String firstChildName = xmlDoc.getDocumentElement().getChildNodes().item(0).getNodeName();
-
-        return firstChildName.equals(XFA_DATA_CHILD_NODE);
-    }
-
-    /**
-     * Add a datasets element to the XML.
-     *
-     * @param xfaData the XML file
-     * @throws Exception a general exception was thrown
-     */
-    private static void addXfaRootHeader(final File xfaData) throws Exception {
-
-        final DocumentBuilderFactory factory = DocumentBuilderFactory.newInstance();
-        final DocumentBuilder builder = factory.newDocumentBuilder();
-
-        final Document oldDoc = builder.parse(xfaData);
-        final Node oldRoot = oldDoc.getDocumentElement();
-        final Document newDoc = builder.newDocument();
-        final Element newRoot = newDoc.createElementNS(XFA_DATA_NS_URI, XFA_DATA_ROOT_NODE);
-
-        newDoc.appendChild(newRoot);
-        final Node dataNode = newRoot;
-        dataNode.appendChild(newDoc.importNode(oldRoot, true));
-
-        final Transformer transformer = TransformerFactory.newInstance().newTransformer();
-        final Result xmlFile = new StreamResult(xfaData);
-        final Source newXml = new DOMSource(newDoc);
-        transformer.transform(newXml, xmlFile);
-    }
-
-    /**
-     * Add a data element to the XML.
-     *
-     * @param xfaData the XML file
-     * @throws Exception a general exception was thrown
-     */
-    private static void addXfaChildHeader(final File xfaData) throws Exception {
-
-        final DocumentBuilderFactory factory = DocumentBuilderFactory.newInstance();
-        final DocumentBuilder builder = factory.newDocumentBuilder();
-
-        final Document oldDoc = builder.parse(xfaData);
-        final Node oldRoot = oldDoc.getElementsByTagName(XFA_DATA_ROOT_NODE).item(0).getFirstChild();
-        final Document newDoc = builder.newDocument();
-        final Element newRoot = newDoc.createElementNS(XFA_DATA_NS_URI, XFA_DATA_ROOT_NODE);
-
-        newDoc.appendChild(newRoot);
-        final Node dataNode = newRoot.appendChild(newDoc.createElement(XFA_DATA_CHILD_NODE));
-        dataNode.appendChild(newDoc.importNode(oldRoot, true));
-
-        final Transformer transformer = TransformerFactory.newInstance().newTransformer();
-        final Result xmlFile = new StreamResult(xfaData);
-        final Source newXml = new DOMSource(newDoc);
-        transformer.transform(newXml, xmlFile);
+        FormImporter.fillXfa(pdfDocument, inputDataUrl, outputUrl);
     }
 
     /**
